Create header file for factoring function. I am still trying to understand how linking and building work. As long as I have the object file for this, then it seems like I can include the header and build a c file that uses it. I have to add the object file at the end of cl command

+++ lib/factoring.c
@@ -8,8 +8,8 @@ This is a very naieve and ineffective way to
 @param factors Pointer to the pointer to the start of the array
 @return num_factors The number of factors (the length of the factors array)
 */
-ULL trialdivision(ULL to_factor, ULL **factors) {
-    ULL num_factors = 0;
+int trialdivision(ULL to_factor, ULL **factors) {
+    int num_factors = 0;
     ULL *factors_ptr = NULL;
 
     for (ULL i = 1; i <= to_factor; i++) {
